+++ pages/courses/[course_id]/topics/[topic_id]/content/[content_id].vue
@@ -24,6 +24,8 @@
 </template>
 
 <script setup lang="ts">
+import { handleContentDone } from "~/utils/content-helpers"
+
 const route = useRoute()
 const isLoading = ref(true)
 const fetchData = async () => {
@@ -36,6 +38,10 @@ const fetchData = async () => {
 
 const fetchedContent = ref()
 
+const isTextContent = (type: string) => {
+  return type === "TEXT"
+}
+
 onMounted(async () => {
   isLoading.value = true
   try {
@@ -43,14 +49,11 @@ onMounted(async () => {
   } finally {
     isLoading.value = false
   }
+
   window.addEventListener("scroll", async (event) => {
     await scrollToBottom(event)
   })
 })
-
-const isTextContent = (type: string) => {
-  return type === "TEXT"
-}
 
 const isPageUnscrollable = () => {
   return (
@@ -61,37 +64,22 @@ const isPageUnscrollable = () => {
 
 onUnmounted(async () => {
   window.removeEventListener("scroll", scrollToBottom)
-  if (isPageUnscrollable()) {
-    await handleContentDone()
+  if (isPageUnscrollable() && fetchedContent.value.type === "TEXT") {
+    await handleContentDone(
+      route.params.content_id as unknown as number,
+      route.params.course_id as unknown as number,
+    )
   }
 })
 
-const scrollToBottom = async (event: Event, content: string = "text") => {
-  if (content === "text") {
+const scrollToBottom = async (event: Event) => {
+  if (fetchedContent.value.type === "TEXT") {
     if (window.innerHeight + window.scrollY >= document.body.offsetHeight) {
-      await handleContentDone()
+      await handleContentDone(
+        route.params.content_id as unknown as number,
+        route.params.course_id as unknown as number,
+      )
     }
-  }
-}
-
-const handleContentDone = async () => {
-  const userStore = useUserStore()
-  const courseProgressStore = useCourseProgressStore()
-  try {
-    const result = await $fetch(`/api/users/${userStore.user?.uid}/progress`, {
-      method: "POST",
-      body: {
-        enrollmentId: courseProgressStore.enrollmentId,
-        contentId: route.params.content_id,
-        userId: userStore.user?.uid,
-      },
-    })
-
-    if (result) {
-      courseProgressStore.updateContentProgress(route.params.course_id)
-    }
-  } catch (error) {
-    console.log(error)
   }
 }
 </script>
